+++ Changelog.md
@@ -124,6 +124,9 @@
     1. Current level is transparent, lower levels opaque, higher levels invisible.
         * [Pull request #1303](https://bitbucket.org/osrf/gazebo/pull-request/1303)
 
+    1. Detach all child manips when item is deleted, added BuildingMaker::DetachAllChildren.
+        * [Pull request #1316](https://bitbucket.org/osrf/gazebo/pull-request/1316)
+
 1. Model editor updates
     1. Fix adding/removing event filters .
         * [Pull request #1279](https://bitbucket.org/osrf/gazebo/pull-request/1279)